make code more modular
