+++ text_from_warc/src/main/java/META-INF/MANIFEST.MF
@@ -1,0 +1,2 @@
+Manifest-Version: 1.0
+Main-Class: warcreader.WarcReader
