+++ text_from_warc/src/main/resources/log4j.properties
@@ -1,9 +1,6 @@
-# Set root logger level to DEBUG and its only appender to A1.
 log4j.rootLogger=ERROR, A1
 
-# A1 is set to be a ConsoleAppender.
 log4j.appender.A1=org.apache.log4j.ConsoleAppender
 
-# A1 uses PatternLayout.
 log4j.appender.A1.layout=org.apache.log4j.PatternLayout
 log4j.appender.A1.layout.ConversionPattern=%-4r [%t] %-5p %c %x - %m%n
